update categoryController delete function

+++ controller/categoryController.js
@@ -1,35 +1,37 @@
 const Category = require('../model/category');
+const Item = require('../model/item');
 const constant = require('../config/constant');
+const async = require('async');
 
 class CategoryController {
 
   getAll(req, res, next) {
     Category.find((err, data) => {
-      if(err) {
+      if (err) {
         next(err);
       }
-      if(!data) {
+      if (!data) {
         res.sendStatus(constant.httpCode.NOT_FOUND);
       }
-      res.status(constant.httpCode.OK).send({category:data});
+      res.status(constant.httpCode.OK).send({category: data});
     })
-    }
+  }
 
   getOne(req, res, next) {
     const categoryId = req.params.id;
     Category.findById(categoryId, (err, data) => {
-      if(err) {
+      if (err) {
         next(err);
       }
-      if(!data) {
+      if (!data) {
         res.sendStatus(constant.httpCode.NOT_FOUND);
       }
-      res.status(constant.httpCode.OK).send({category:data});
+      res.status(constant.httpCode.OK).send({category: data});
     })
   }
 
   create(req, res, next) {
-     Category.create(req.body, (err, data) => {
+    Category.create(req.body, (err, data) => {
       if (err) {
         next(err);
       }
@@ -39,25 +41,45 @@ class CategoryController {
 
   delete(req, res, next) {
     const categoryId = req.params.id;
-    Category.findByIdAndRemove(categoryId, (err, data) => {
+    async.waterfall([
+      (done) => {
+        Item.findOne({categoryId}, done);
+      },
+      (data, done) => {
+        if (data) {
+          done(true, null);
+        } else {
+          Category.findByIdAndRemove(categoryId, (err, data) => {
+            if (!data) {
+              done(false, null);
+            }
+            done(err, data);
+          });
+        }
+      }
+    ], (err)=> {
+      if (err === true) {
+        res.sendStatus(constant.httpCode.BAD_REQUEST);
+      }
+      if (err === false) {
+        res.sendStatus(constant.httpCode.NOT_FOUND);
+      }
       if (err) {
         next(err);
-      }
-      if(!data) {
-        res.sendStatus(constant.httpCode.NOT_FOUND);
       }
       res.sendStatus(constant.httpCode.NO_CONTENT);
     });
 
+
   }
 
-  update(req, res, next){
+  update(req, res, next) {
     const categoryId = req.params.id;
     Category.findByIdAndUpdate(categoryId, req.body, (err, data) => {
       if (err) {
         next(err);
       }
-      if(!data) {
+      if (!data) {
         res.sendStatus(constant.httpCode.NOT_FOUND);
       }
       res.sendStatus(constant.httpCode.NO_CONTENT);
